refactor: refactored llama specific functions to a separate folder
